Allow Adapters to be extended - Changed private methods to protected - Marked all methods to be virtual
Updated ObjectAdapter to allow for typed ParsedPath and ObjectVisitor classes
- Added Interfaces for ParsedPath and ObjectVisitor
- Created a new Generic ObjectAdapter which allows the implementations
of ParsedPath and ObjectVisitor to be set
- Updated ObjectAdapter to inherit from the Generic ObjectAdapter using
the default implementations of ParsedPath and ObjectVisitor
