+++ src/Microsoft.AspNetCore.JsonPatch/Internal/IObjectVisitor.cs
@@ -1,0 +1,7 @@
+﻿namespace Microsoft.AspNetCore.JsonPatch.Internal
+{
+    public interface IObjectVisitor
+    {
+        bool TryVisit(ref object target, out IAdapter adapter, out string errorMessage);
+    }
+}
