+++ src/Microsoft.AspNetCore.JsonPatch/Internal/IParsedPath.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace Microsoft.AspNetCore.JsonPatch.Internal
+{
+    public interface IParsedPath
+    {
+        string LastSegment { get; }
+        IReadOnlyList<string> Segments { get; }
+    }
+}
