+++ src/Microsoft.AspNetCore.JsonPatch/Internal/ParsedPath.cs
@@ -8,22 +8,16 @@ using System.Text;
 
 namespace Microsoft.AspNetCore.JsonPatch.Internal
 {
-    public struct ParsedPath
+    public class ParsedPath : IParsedPath
     {
         private static readonly string[] Empty = null;
 
         private readonly string[] _segments;
 
-        private static List<char> _separators = new List<char> { '/' };
+        public static List<char> Separators = new List<char> { '/' };
 
-        public ParsedPath(string path)
+        public ParsedPath(string path):this(path, null)
         {
-            if (path == null)
-            {
-                throw new ArgumentNullException(nameof(path));
-            }
-
-            _segments = ParsePath(path);
         }
 
         public ParsedPath(string path, List<char> separators)
@@ -32,7 +26,10 @@ namespace Microsoft.AspNetCore.JsonPatch.Internal
             {
                 throw new ArgumentNullException(nameof(path));
             }
-            _separators = separators;
+            if (separators!=null)
+            {
+                Separators = separators;
+            }
 
             _segments = ParsePath(path);
         }
@@ -59,7 +56,7 @@ namespace Microsoft.AspNetCore.JsonPatch.Internal
 
             for (var i = 0; i < path.Length; i++)
             {
-                if (_separators.Contains(path[i]))
+                if (Separators.Contains(path[i]))
                 {
                     if (sb.Length > 0)
                     {
